refactor(Tooltip): move Tooltip to future structure
