+++ packages/components/src/Tooltip/Tooltip.module.scss
@@ -3,7 +3,7 @@
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/animation";
 @import "~@kaizen/design-tokens/sass/typography";
-@import "../../../../styles/utils/layers";
+@import "../../styles/utils/layers";
 @import "./variables";
 
 .tooltip {
