+++ packages/components/src/Tooltip/index.ts
@@ -1,0 +1,1 @@
+export * from "./Tooltip"
