+++ packages/components/src/__future__/Tooltip/_docs/Tooltip.spec.stories.tsx
@@ -6,8 +6,8 @@ import { AddIcon } from "~components/Icon"
 import { Tab, TabList, TabPanel, TabPanels, Tabs } from "~components/Tabs"
 import { Text } from "~components/Text"
 import { Button, IconButton } from "~components/__actions__/v2"
+import { Focusable } from "~components/__future__"
 import { Tag } from "~components/__future__/Tag"
-import { Focusable } from "~components/__overlays__/v3"
 import { Tooltip, TooltipTrigger } from "../index"
 
 const meta = {
