+++ packages/components/src/__future__/Tooltip/_docs/Tooltip.stories.tsx
@@ -2,8 +2,7 @@ import React, { FunctionComponent } from "react"
 import { Meta, StoryObj } from "@storybook/react"
 import isChromatic from "chromatic"
 import { Button } from "~components/__actions__/v2"
-import { Tag } from "~components/__future__"
-import { Focusable } from "~components/__overlays__/v3"
+import { Tag, Focusable } from "~components/__future__"
 import { Tooltip, TooltipTrigger } from "../index"
 import * as testStories from "./Tooltip.spec.stories"
