+++ packages/components/src/__future__/Tooltip/_docs/assets/tooltip_placement.png
