+++ packages/components/src/__future__/Tooltip/index.ts
@@ -1,2 +1,3 @@
+export * from "./subcomponents/Focusable"
 export * from "./Tooltip"
 export * from "./TooltipTrigger"
