+++ packages/components/src/__future__/index.ts
@@ -1,3 +1,4 @@
 export * from "./Select"
 export * from "./Tag"
+export * from "./Tooltip"
 export * from "../__layout__/Workflow/v2"
