--- packages/components/src/__overlays__/v3.ts
@@ -1,2 +1,0 @@
-export * from "./Tooltip/v3"
-export * from "./Focusable/v3"
