+++ packages/components/src/index.ts
@@ -54,10 +54,10 @@ export * from "./Tile"
 export * from "./TimeField"
 export * from "./TitleBlockZen"
 export * from "./ToggleSwitch"
+export * from "./Tooltip"
 export * from "./utils"
 export * from "./VisuallyHidden"
 export * from "./Well"
 export * from "./__layout__/Workflow/v1"
-export * from "./__overlays__/Tooltip/v1"
 export * from "./__actions__/Menu/v1"
 export * from "./__actions__/Button/v1"
